mise en place de la fonction retour avec suppression de la photo

+++ Frontend/src/app/pages/page-container/parcour/experiences/experience/experience.component.ts
@@ -51,7 +51,9 @@ export class ExperienceComponent implements OnInit {
   }
 
   deleteLinkImage(experienceImage: String){
+    if(experienceImage !== null){
     this.uploadService.removeFileOfCard(experienceImage.split('/')[3]);
+    }
   }
 
   ngOnDestroy(): void {
